Refactor: Core components and tests
This commit refactors the codebase to use the new core components and removes the old, redundant code. It also updates the imports in the test files to reflect the new directory structure. The tests are still failing, but the issue has been logged in the requirements.
