+++ tests/conftest.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 
 # Add the custom_components directory to the Python path
-sys.path.insert(0, str(Path(__file__).parent.parent / "custom_components"))
+sys.path.insert(0, str(Path(__file__).parent.parent))
 
 
 @pytest.fixture(autouse=True)
