+++ tests/sensor/test_network_clients.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock
 
 from homeassistant.core import HomeAssistant
 
-from custom_components.meraki_ha.sensor.network_clients import (
+from custom_components.meraki_ha.core.sensor.network_clients import (
     MerakiNetworkClientsSensor,
 )
