+++ tests/sensor/test_radio_settings.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock
 
 from homeassistant.core import HomeAssistant
 
-from custom_components.meraki_ha.sensor.radio_settings import (
+from custom_components.meraki_ha.core.sensor.radio_settings import (
     MerakiRadioSettingsSensor,
 )
