+++ tests/switch/test_meraki_ssid_device_switch.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock
 
 from homeassistant.core import HomeAssistant
 
-from custom_components.meraki_ha.switch.meraki_ssid_device_switch import (
+from custom_components.meraki_ha.core.switch.meraki_ssid_device_switch import (
     MerakiSSIDEnabledSwitch,
     MerakiSSIDBroadcastSwitch,
 )
